enum package moved to model package
