+++ main/src/main/kotlin/com/ip13/main/model/dto/RestaurantAddTicketResultDto.kt
@@ -1,6 +1,6 @@
 package com.ip13.main.model.dto
 
-import com.ip13.main.model.entity.enums.RestaurantAddStatus
+import com.ip13.main.model.enums.RestaurantAddStatus
 
 data class RestaurantAddTicketResultDto(
     val restaurantAddTicketId: Int = 0,
