+++ main/src/test/kotlin/com/ip13/main/controller/AdminControllerTest.kt
@@ -1,6 +1,6 @@
 package com.ip13.main.controller
 
-import com.ip13.main.model.entity.enums.Role
+import com.ip13.main.model.enums.Role
 import com.ip13.main.security.entity.User
 import com.ip13.main.security.service.TokenService
 import com.ip13.main.security.service.UserService
